refactor(#58): move page containers to pages subfolder - link to policy instaed of about in footer menu
