+++ src/containers/AppRoutes/AppRoutes.tsx
@@ -2,21 +2,19 @@ import * as React from 'react'
 import {Route} from 'react-router-dom'
 
 import {PrivateRoute} from './PrivateRoute'
-import IntentionPage from '../IntentionPage'
-import IntentionList from '../IntentionList'
-import LoginPage from '../LoginPage'
-import {AddIntentionPage} from '../AddIntentionPage'
-import {HowItWorks} from 'src/components/HowItWorks'
-import {About} from 'src/components/About'
-import {PrivacyPolicy} from 'src/components/PrivacyPolicy'
+import IntentionPage from 'src/pages/IntentionPage'
+import IntentionList from 'src/pages/IntentionList'
+import LoginPage from 'src/pages/LoginPage'
+import {AddIntentionPage} from 'src/pages/AddIntentionPage'
+import {HowItWorks} from 'src/pages/HowItWorks'
+import {About} from 'src/pages/About'
 
 export const AppRoutes = () => {
   // TODO: #37 Add "how it works" on homepage
   return (
     <>
       <Route path="/" exact={true} component={IntentionList} />
-      <Route path="/about" exact={true} component={About} />
-      <Route path="/policy" exact={true} component={PrivacyPolicy} />
+      <Route path="/policy" exact={true} component={About} />
       <Route path="/login" exact={true} component={LoginPage} />
       <Route path="/how-it-works" exact={true} component={HowItWorks} />
       <PrivateRoute path="/add-intention">
