+++ src/pages/HowItWorks/HowItWorks.tsx
