+++ src/pages/HowItWorks/index.tsx
